backend/crawler/pyproject.toml
[project]
name = "crawler"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.10.0",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.4.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "redis>=5.2.0",
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "pydantic-settings>=2.14.0",
    "newspaper4k>=0.9.3",
    "readability-lxml>=0.8.1",
    "lxml>=5.3.0",
]
